+++ func/notify.js
@@ -15,35 +15,27 @@ module.exports = {
 				const messageData = [];
 				const removedArgs = args.filter((v) => !checkedArgs.includes(v));
 				if (removedArgs.length) messageData.push(`The following bosses were found in the saved list. They won't be added: \`${removedArgs.join("`, `")}\``);
-				pokeNavCheck(checkedArgs, message).then(async ([result, m]) => {
-					m.forEach((item) => messageData.push(item));
+				pokeNavCheck(checkedArgs, message).then(async ([result, md]) => {
+					md.forEach((item) => messageData.push(item));
 					if (typeof result == "string") return reject([result, messageData]);
 					roleMake(result, message).then(async () => {
-						const notifyChannel = await message.guild.channels.fetch(ops.notifyReactionChannel);
-						console.log("Updating saved list");
+						const newList = new Discord.Collection;
 						list.forEach((arr, key) => {
 							if (result.has(key)) {
 								const resultArr = result.get(key);
-								const newArr = arr.concat(resultArr);
-								list.set(key, newArr);
+								const newArr = arr.concat(resultArr).sort();
+								newList.set(key, newArr);
 							}
 						});
 						result.forEach((arr, key) => {
 							if (!list.has(key)) {
-								list.set(key, arr);
+								newList.set(key, arr.sort());
 							}
-						}); // exists roles zapdos
-						module.exports.saveNotifyList();
-						// const row = new Discord.MessageActionRow()
-						// .addComponents([
-						// 	new Discord.MessageSelectMenu()
-						// 	.setCustomId()
-						// ])
-						// notifyChannel.send();
-
-
-						// message.reply(messageData.join("\n"));
-						// resolve();
+						});
+						module.exports.deleteAndMakeMessages(message, newList, messageData).catch((err) => {
+							message.reply(err);
+							console.error(err);
+						});
 					});
 				});
 			});
@@ -108,6 +100,64 @@ module.exports = {
 			});
 		});
 	},
+	checkButtonInput(interaction){
+		return new Promise((resolve) => {
+			const val = interaction.customId;
+			if (list.some((arr) => arr.includes(val))) {
+				resolve(true);
+				buttonInput(interaction);
+			} else resolve(false);
+		});
+	},
+	async deleteAndMakeMessages(input, newList, messageData) {
+		let notifyChannel;
+		if (input instanceof Discord.Message) notifyChannel = await input.guild.channels.fetch(ops.notifyReactionChannel);
+		else if (input instanceof Discord.Guild) notifyChannel = await input.channels.fetch(ops.notifyReactionChannel);
+		else throw "Could not load new notification reaction message";
+		if (!newList) newList = list;
+		const existingButtons = await notifyChannel.messages.fetch({ limit: 6 }).then((ms) => ms.filter((msg) => !msg.pinned));
+		notifyChannel.bulkDelete(existingButtons).then(() => {
+			console.log("Constructing buttons");
+			newList.forEach(async (arr, key) => {
+				if (arr.length > 25) {
+					throw "I cannot currently process more than 25 bosses per tier. Please clear before trying again.";
+				}
+				const embed = new Discord.MessageEmbed()
+				.setTitle(key)
+				.setDescription(`Click on a ${key} to be notified when a new raid is posted.\nClick it again to remove the notification.`);
+				let buttonArr = [];
+				const rowArr = [];
+				for (const boss of arr) {
+					if (buttonArr.length == 5) {
+						rowArr.push(new Discord.MessageActionRow()
+						.addComponents(buttonArr));
+						buttonArr = [];
+					}
+					const bossName = boss[0].toUpperCase() + boss.substring(1);
+					const button = new Discord.MessageButton()
+					.setCustomId(boss)
+					.setLabel(bossName)
+					.setStyle("PRIMARY");
+					buttonArr.push(button);
+					if (arr.indexOf(boss) == arr.length - 1) {
+						rowArr.push(new Discord.MessageActionRow()
+						.addComponents(buttonArr));
+					}
+				}
+				console.log(`Sending ${key} button message`);
+				notifyChannel.send({ components: rowArr, embeds: [embed] }).then(() => {
+					if (newList.lastKey() == key) {
+						list = newList;
+						console.log("Updating saved list");
+						module.exports.saveNotifyList().then(() => {
+							if (input instanceof Discord.Message) input.reply(`Notifications added.\n\nErrors:\n• ${messageData.join("\n• ")}`);
+							return;
+						});
+					}
+				});
+			});
+		}).catch((err) => console.error(err));
+	},
 };
 
 async function argsCheck(args) {
@@ -137,6 +187,7 @@ async function pokeNavCheck(data, message, messageData, i, result) {
 					const respTitle = resp.first().embeds[0].title;
 					pokenavChannel.bulkDelete(2).catch(() => console.error("Could not delete a message in the pokenavChannel"));
 					if (respTitle == "Error") {
+						console.log(`${mon} was not found by pokenav.`);
 						messageData.push(`PokeNav could not find \`${mon}\`. Please try again for that boss.`);
 					} else {
 						const tierLocation = respTitle.toLowerCase().indexOf("tier");
@@ -160,6 +211,7 @@ async function pokeNavCheck(data, message, messageData, i, result) {
 					return console.error("An unexpected error in ]notify. error:", e);
 				}
 			}).catch(() => {
+				console.log(`${mon} took more than 20 seconds for pokenav to find...?`);
 				messageData.push(`PokeNav did not respond quickly enough for \`${mon}\`. Please try again for that boss.`);
 				if (data.indexOf(mon) == data.length - 1) {
 					if (result.size > 0) {
@@ -203,4 +255,21 @@ function roleMake(input, message) {
 			}
 		}
   });
+}
+
+async function buttonInput(interaction){
+	const val = interaction.customId;
+	const member = await interaction.member.fetch();
+	const role = await interaction.guild.roles.fetch().then(roles => roles.find(r => r.name == val));
+	if (!member.roles.cache.has(role.id)) {
+		member.roles.add(role).then(() => {
+			interaction.reply({ content: `Notifications activated for: \`${val}\`.`, ephemeral:true });
+			return;
+		});
+	} else {
+		member.roles.remove(role).then(() => {
+			interaction.reply({ content: `Notifications **disabled** for: \`${val}\`.`, ephemeral:true });
+			return;
+		});
+	}
 }