+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v1.3 (2022-Jul-07)
+* Made a raid announce channel category switcher for PokeNav
+ * Use `arc` and `rrc` to link categories to announce channels, and when they reach **catLimit** channels, it will swap to another linked category, so long as that category has less than half of **catLimit** channels
+* Added `]notify`
+
 # v1.2 (2022-Jun-06)
 * Updated to where the main bot's cleanup process was at:
  * Made the bot greedy, deleting *everything* that isn't a raid.